feat(terminal): origin stack; local coordinates in write, read and cursor calls
push_origin/pop_origin accumulate the owners' origins; write_cell,
write_line, read_cell, show_cursor and push_clip take local i16
coordinates and translate. With an empty stack behaviour is unchanged.
TestBackend records the last cursor position. Step 1 of
docs/OWNER-COORDINATES.md.

+++ src/views/dir_listbox.rs
@@ -340,8 +340,8 @@ impl View for DirListBox {
 
             for (x, ch) in padded.chars().take(width).enumerate() {
                 terminal.write_cell(
-                    (self.core.bounds.a.x + x as i16) as u16,
-                    (self.core.bounds.a.y + y as i16) as u16,
+                    self.core.bounds.a.x + x as i16,
+                    self.core.bounds.a.y + y as i16,
                     crate::core::draw::Cell::new(ch, color),
                 );
             }
